Extract out UsageStore interface from SpaceCheckingKeyFilter
